requirements.txt
gunicorn
line-bot-sdk==2.3.0
flask==2.1.2
Werkzeug==2.2.2
requests==2.28.0